ml-model/requirements.txt
# Flight Schedule Optimization ML Model Dependencies

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.13.0

# Machine Learning
scikit-learn>=1.3.0
networkx>=3.0
joblib>=1.3.0

# API Framework
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6

# LLM Integration
ollama>=0.5.0
httpx>=0.27.0

# Data validation and type hints
pydantic>=2.0.0
typing-extensions>=4.0.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
alembic>=1.12.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3

# Reporting / PDF
reportlab>=4.0.0
